geoviality-ia/requirements.txt
requests
python-dotenv
pymongo
ultralytics==8.2.79
opencv-python
torch==2.3.1
pandas
pyngrok
pika
pydantic
fastapi
pytz